--- packages/frontend/src/pages/Swap/components/SwapMinerTip/slider.styles.css
@@ -1,3 +1,0 @@
-.rc-slider-mark-text {
-  text-transform: capitalize;
-}
